mcp_servers/newspaper_server/pyproject.toml
[project]
name = "newspaper_server"
version = "0.1.0"
description = "MCP Server for newspaper_server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lxml",
    "lxml_html_clean",
    "mcp",
    "newspaper3k",
    "nltk",
    "pydantic",
    "python-dotenv",
    "structlog",
    "wrapt",
]

[tool.uv]
dev-dependencies = []
